# apps/api/generated/openapi.generated.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"TravelWindowKindDto": {
				"type": "string",
				"enum": [
					"vacation",
					"work"
				]
			},
			"TravelWindowDto": {
				"properties": {
					"accountName": {
						"type": "string",
						"nullable": true
					},
					"accountId": {
						"type": "string",
						"nullable": true
					},
					"endDate": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"kind": {
						"$ref": "#/components/schemas/TravelWindowKindDto"
					},
					"name": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"accountName",
					"accountId",
					"endDate",
					"startDate",
					"kind",
					"name",
					"id"
				],
				"type": "object"
			},
			"TravelWindowsDto": {
				"properties": {
					"windows": {
						"items": {
							"$ref": "#/components/schemas/TravelWindowDto"
						},
						"type": "array"
					}
				},
				"required": [
					"windows"
				],
				"type": "object"
			},
			"TravelWindowWriteDto": {
				"properties": {
					"accountName": {
						"type": "string",
						"nullable": true
					},
					"accountId": {
						"type": "string",
						"nullable": true
					},
					"endDate": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"kind": {
						"$ref": "#/components/schemas/TravelWindowKindDto"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"accountName",
					"accountId",
					"endDate",
					"startDate",
					"kind",
					"name"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@budget-tools/api",
		"version": "0.0.0",
		"contact": {}
	},
	"paths": {
		"/travel-windows": {
			"get": {
				"operationId": "ListTravelWindows",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TravelWindowsDto"
								}
							}
						}
					}
				},
				"summary": "listTravelWindows",
				"tags": [
					"travel-windows"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTravelWindow",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TravelWindowDto"
								}
							}
						}
					},
					"409": {
						"description": "Overlapping travel window"
					}
				},
				"summary": "createTravelWindow",
				"tags": [
					"travel-windows"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TravelWindowWriteDto"
							}
						}
					}
				}
			}
		},
		"/travel-windows/{id}": {
			"put": {
				"operationId": "UpdateTravelWindow",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TravelWindowDto"
								}
							}
						}
					},
					"409": {
						"description": "Overlapping travel window"
					}
				},
				"summary": "updateTravelWindow",
				"tags": [
					"travel-windows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TravelWindowWriteDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTravelWindow",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "deleteTravelWindow",
				"tags": [
					"travel-windows"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}